+++ helixtest/crates/crypt4gh-tests/src/tests.rs
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
 use anyhow::Result;
 use common::crypto::{corrupt_file, decrypt_file, decrypt_partial, encrypt_file};
 use common::util::sha256_file;